chore: Unified sharing state unit test file
